requirements.txt
spacy
transformers
pdfminer.six
docx2txt
nltk
scikit-learn
gensim
numpy
pandas
